# checksum: using xxhash64() & MAX_INT is 50-80% faster than xxhash32