--- a/AUTHORS
+++ b/AUTHORS
@@ -1,0 +1,2 @@
+Alexander Gallego
+Noah Watkins