--- a/ci/local.sh
+++ b/ci/local.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+
+# Copyright 2018 SMF Authors
+#
+
 set -e
 set -x