--- a/src/benchmarks/CMakeLists.txt
+++ b/src/benchmarks/CMakeLists.txt
@@ -14,3 +14,10 @@ smf_test(
   SOURCE_DIRECTORY ${BENCH_ROOT}/core_balance
   LIBRARIES benchmark smf_hashing fmt seastar
   )
+smf_test(
+  BENCHMARK_TEST
+  BINARY_NAME checksum
+  SOURCES ${BENCH_ROOT}/checksum_bench/main.cc
+  SOURCE_DIRECTORY ${BENCH_ROOT}/checksum_bench
+  LIBRARIES benchmark smf_hashing fmt seastar
+  )