--- a/src/rpc/rpc_connection_limits.cc
+++ b/src/rpc/rpc_connection_limits.cc
@@ -23,32 +23,6 @@ rpc_connection_limits::estimate_request_size(uint64_t serialized_size) {
   return (basic_request_size + serialized_size) * bloat_factor;
 }
 
-seastar::future<>
-rpc_connection_limits::wait_for_resources(uint64_t memory_consumed) {
-  LOG_ERROR_IF(
-    memory_consumed > max_memory,
-    "memory to serve request `{}`, exceeds max available memory `{}`",
-    memory_consumed, max_memory);
-
-  return resources_available.wait(memory_consumed);
-}
-
-void
-rpc_connection_limits::release_resources(uint64_t memory_consumed) {
-  resources_available.signal(memory_consumed);
-}
-void
-rpc_connection_limits::release_payload_resources(uint64_t payload_size) {
-  release_resources(estimate_request_size(payload_size));
-}
-
-seastar::future<>
-rpc_connection_limits::wait_for_payload_resources(uint64_t payload_size) {
-  return wait_for_resources(estimate_request_size(payload_size));
-}
-
-rpc_connection_limits::~rpc_connection_limits() {}
-
 std::ostream &
 operator<<(std::ostream &o, const rpc_connection_limits &l) {
   o << "rpc_connection_limits{'basic_req_size':"