--- a/src/rpc/rpc_header_utils.h
+++ b/src/rpc/rpc_header_utils.h
@@ -7,10 +7,17 @@
 #include "rpc/rpc_generated.h"
 
 namespace smf {
+
+inline static uint32_t
+rpc_checksum_payload(const char *payload, uint32_t size) {
+  return std::numeric_limits<uint32_t>::max() & xxhash_64(payload, size);
+}
+
 template <typename T>
 inline void
 checksum_rpc(T &hdr, const char *payload, uint32_t size) {
-  hdr.mutate_checksum(xxhash_32(payload, size));
+  hdr.mutate_checksum(rpc_checksum_payload(payload, size));
   hdr.mutate_size(size);
 }
+
 }  // namespace smf