--- a/src/rpc/rpc_recv_context.cc
+++ b/src/rpc/rpc_recv_context.cc
@@ -4,9 +4,9 @@
 
 #include <flatbuffers/minireflect.h>
 
-#include "rpc/rpc_generated.h"
-#include "hashing/hashing_utils.h"
 #include "platform/log.h"
+#include "rpc/rpc_generated.h"
+#include "rpc/rpc_header_utils.h"
 
 namespace std {
 ostream &
@@ -19,7 +19,6 @@ operator<<(ostream &o, const smf::rpc::header &h) {
 }  // namespace std
 
 namespace smf {
-namespace exp = std::experimental;
 
 rpc_recv_context::rpc_recv_context(rpc::header                     hdr,
                                    seastar::temporary_buffer<char> body)
@@ -45,35 +44,21 @@ rpc_recv_context::status() const {
 
 seastar::future<seastar::temporary_buffer<char>>
 read_payload(rpc_connection *conn, size_t payload_size) {
-  if (!conn->limits) {
-    return conn->istream.read_exactly(payload_size);
-  } else {
-    return conn->limits->wait_for_payload_resources(payload_size)
-      .then([payload_size, conn]() {
-        // set timeout
-        seastar::timer<> body_timeout;
-        body_timeout.set_callback(
-          [ max_timeout = conn->limits->max_body_parsing_duration, conn ] {
-            LOG_ERROR(
-              "Parsing the body of the connnection exceeded max_timeout: {}ms",
-              std::chrono::duration_cast<std::chrono::milliseconds>(max_timeout)
-                .count());
-            conn->set_error("Connection body parsing exceeded timeout");
-            conn->socket.shutdown_input();
-          });
-
-        // TODO(agallego) - this should be a multiplier on the size
-        // for example 100ms * byte up to a max of 10mins
-
-        body_timeout.arm(conn->limits->max_body_parsing_duration);
-        return conn->istream.read_exactly(payload_size)
-          .then([body_timeout = std::move(body_timeout)](auto payload) mutable {
-            body_timeout.cancel();
-            return seastar::make_ready_future<decltype(payload)>(
-              std::move(payload));
-          });
-      });
-  }
+  auto             timeout = conn->limits->max_body_parsing_duration;
+  seastar::timer<> body_timeout;
+  body_timeout.set_callback([timeout, conn] {
+    LOG_ERROR(
+      "Parsing the body of the connnection exceeded max_timeout: {}ms",
+      std::chrono::duration_cast<std::chrono::milliseconds>(timeout).count());
+    conn->set_error("Connection body parsing exceeded timeout");
+    conn->socket.shutdown_input();
+  });
+  body_timeout.arm(conn->limits->max_body_parsing_duration);
+  return conn->istream.read_exactly(payload_size)
+    .then([body_timeout = std::move(body_timeout)](auto payload) mutable {
+      body_timeout.cancel();
+      return seastar::make_ready_future<decltype(payload)>(std::move(payload));
+    });
 }
 
 constexpr uint32_t
@@ -85,42 +70,42 @@ max_flatbuffers_size() {
   return static_cast<uint32_t>(FLATBUFFERS_MAX_BUFFER_SIZE);
 }
 
-seastar::future<exp::optional<rpc_recv_context>>
-process_payload(rpc_connection *conn, rpc::header hdr) {
-  using ret_type = exp::optional<rpc_recv_context>;
+seastar::future<stdx::optional<rpc_recv_context>>
+rpc_recv_context::parse_payload(rpc_connection *conn, rpc::header hdr) {
+  using ret_type = stdx::optional<rpc_recv_context>;
   return read_payload(conn, hdr.size())
     .then([hdr](seastar::temporary_buffer<char> body) mutable {
       if (hdr.size() != body.size()) {
         LOG_ERROR("Read incorrect number of bytes `{}`, expected header: `{}`",
                   body.size(), hdr);
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
       if (hdr.size() > max_flatbuffers_size()) {
         LOG_ERROR("Bad payload. Body is >  FLATBUFFERS_MAX_BUFFER_SIZE");
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
       if (hdr.bitflags() &
           rpc::header_bit_flags::header_bit_flags_has_payload_headers) {
         LOG_ERROR("Reading payload headers is not yet implemented");
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
 
-      const uint32_t xx = xxhash_32(body.get(), body.size());
+      const uint32_t xx = rpc_checksum_payload(body.get(), body.size());
       if (xx != hdr.checksum()) {
         LOG_ERROR("Payload checksum `{}` does not match header checksum `{}`",
                   xx, hdr.checksum());
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
 
       rpc_recv_context ctx(hdr, std::move(body));
       return seastar::make_ready_future<ret_type>(
-        exp::optional<rpc_recv_context>(std::move(ctx)));
+        stdx::optional<rpc_recv_context>(std::move(ctx)));
     });
 }
 
-seastar::future<exp::optional<rpc_recv_context>>
-rpc_recv_context::parse(rpc_connection *conn) {
-  using ret_type = exp::optional<rpc_recv_context>;
+seastar::future<stdx::optional<rpc::header>>
+rpc_recv_context::parse_header(rpc_connection *conn) {
+  using ret_type = stdx::optional<rpc::header>;
 
   static constexpr size_t kRPCHeaderSize = sizeof(rpc::header);
   DLOG_THROW_IF(
@@ -134,16 +119,15 @@ rpc_recv_context::parse(rpc_connection *conn) {
         LOG_ERROR_IF(conn->is_valid(),
                      "Invalid header size `{}`, expected `{}`, skipping req",
                      header.size(), kRPCHeaderSize);
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
       rpc::header hdr;
       std::memcpy(&hdr, header.get(), header.size());
-
       if (hdr.size() == 0) {
         LOG_ERROR("Emty body to parse. skipping");
-        return seastar::make_ready_future<ret_type>(exp::nullopt);
+        return seastar::make_ready_future<ret_type>(stdx::nullopt);
       }
-      return process_payload(conn, hdr);
+      return seastar::make_ready_future<ret_type>(std::move(hdr));
     })
     .finally([conn] { conn->istream_active_parser--; });
 }