Details:
```
---------------------------------------------------------------------
BM_hash32/65536/65536               8265 ns       8259 ns      83232
BM_hash32/1048576/1048576         126584 ns     126482 ns       5489
BM_hash32/536870912/536870912   67708174 ns   67540241 ns         10
BM_hash64/65536/65536               4413 ns       4411 ns     158959
BM_hash64/1048576/1048576          63571 ns      63555 ns      10846
BM_hash64/536870912/536870912   37536790 ns   37418839 ns         19
```