@senior7515 2018-02-27
I think i found a deadlock...


```
# SERVER

./src/demo_apps/demo_server  -m 4G --cpuset 0-3 --poll-mode

```


```
# CLIENT

./src/demo_apps/demo_client  --cpuset 6 -m 4G --poll-mode --req-num 1000000
```

Never finishes... :'( 

Debugging. 

Also happens on master. doing a  `perf top` looks like it's stuck in a seastar follin function... Going to test master to see if it happens. I know duarte pushed a new series of updates for this exact set of conditions.