@senior7515 2018-02-28
This solves a lot of issues! with benchmarking. Handles semaphore asignment correctly.